refactor: Cleanup and improve state management logic in reader
Remove the existing reader bottom sheet menu for now, to add a better option later
