+++ app/src/main/java/com/starry/myne/ui/screens/reader/main/viewmodel/ReaderViewModel.kt
@@ -15,39 +15,46 @@
  */
 
 
-package com.starry.myne.ui.screens.reader.viewmodels
+package com.starry.myne.ui.screens.reader.main.viewmodel
 
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.setValue
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.starry.myne.database.library.LibraryDao
 import com.starry.myne.database.reader.ReaderDao
 import com.starry.myne.database.reader.ReaderData
 import com.starry.myne.epub.EpubParser
-import com.starry.myne.epub.models.EpubBook
 import com.starry.myne.epub.models.EpubChapter
+import com.starry.myne.epub.models.EpubImage
 import com.starry.myne.helpers.PreferenceUtil
-import com.starry.myne.ui.screens.reader.others.ReaderFont
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.StateFlow
+import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.launch
 import java.io.FileInputStream
 import javax.inject.Inject
 
 
 data class ReaderScreenState(
+    // Screen state
     val isLoading: Boolean = true,
     val shouldShowLoader: Boolean = false,
     val showReaderMenu: Boolean = false,
-    val fontSize: Int = 18,
+    val currentChapterIndex: Int = 0,
+    val currentChapter: EpubChapter = EpubChapter("", "", "", ""),
+    val chapterScrollPercent: Float = 0f,
+    // Book data
+    val title: String = "",
+    val chapters: List<EpubChapter> = emptyList(),
+    val images: List<EpubImage> = emptyList(),
+    // Reader data
+    val hasProgressSaved: Boolean = false,
+    val lastChapterIndex: Int = 0,
+    val lastChapterOffset: Int = 0,
+    // Typography
+    val fontSize: Int = 100,
     val fontFamily: ReaderFont = ReaderFont.System,
-    val epubBook: EpubBook? = null,
-    val readerData: ReaderData? = null
 )
 
 @HiltViewModel
@@ -58,28 +65,30 @@ class ReaderViewModel @Inject constructor(
     private val epubParser: EpubParser
 ) : ViewModel() {
 
-    var state by mutableStateOf(
-        ReaderScreenState(
-            fontFamily = getFontFamily(),
-            fontSize = getFontSize()
-        )
-    )
-
-    private val _chapterScrolledPercent = MutableStateFlow(0f)
-    val chapterScrolledPercent: StateFlow<Float> = _chapterScrolledPercent
-
-    private val _visibleChapterIndex = MutableStateFlow(0)
-    val visibleChapterIndex: StateFlow<Int> = _visibleChapterIndex
-
-    private val _currentChapter = MutableStateFlow(
-        state.epubBook?.chapters?.getOrNull(_visibleChapterIndex.value)
-    )
-    val currentChapter: StateFlow<EpubChapter?> = _currentChapter
+    // Mutable state flow to update the state.
+    private val _state = MutableStateFlow(ReaderScreenState())
+    val state = _state.asStateFlow()
 
     init {
         viewModelScope.launch {
-            _visibleChapterIndex.collect { index ->
-                _currentChapter.value = state.epubBook?.chapters?.getOrNull(index)
+            _state.value = _state.value.copy(
+                fontSize = preferenceUtil.getInt(PreferenceUtil.READER_FONT_SIZE_INT, 100)
+            )
+            _state.value = _state.value.copy(
+                fontFamily = ReaderFont.getFontById(
+                    preferenceUtil.getString(
+                        PreferenceUtil.READER_FONT_STYLE_STR,
+                        ReaderFont.System.id
+                    )!!
+                )
+            )
+            // Collect the state to update the current chapter.
+            _state.collect {
+                if (state.value.isLoading) return@collect
+                if (state.value.chapters.isEmpty()) return@collect
+                _state.value = _state.value.copy(
+                    currentChapter = _state.value.chapters[_state.value.currentChapterIndex]
+                )
             }
         }
     }
@@ -87,7 +96,9 @@ class ReaderViewModel @Inject constructor(
     fun loadEpubBook(libraryItemId: Int, onLoaded: (ReaderScreenState) -> Unit) {
         viewModelScope.launch(Dispatchers.IO) {
             val libraryItem = libraryDao.getItemById(libraryItemId)
-            state = state.copy(shouldShowLoader = !epubParser.isBookCached(libraryItem!!.filePath))
+            _state.value = _state.value.copy(
+                shouldShowLoader = !epubParser.isBookCached(libraryItem!!.filePath)
+            )
             val readerData = readerDao.getReaderData(libraryItemId)
             // parse and create epub book
             var epubBook = epubParser.createEpubBook(libraryItem.filePath)
@@ -97,13 +108,24 @@ class ReaderViewModel @Inject constructor(
             if (epubBook.language == "zh" && !libraryItem.isExternalBook) {
                 epubBook = epubParser.createEpubBook(libraryItem.filePath, shouldUseToc = false)
             }
-            state = state.copy(epubBook = epubBook, readerData = readerData)
-            onLoaded(state)
+            _state.value = _state.value.copy(
+                title = libraryItem.title,
+                chapters = epubBook.chapters,
+                images = epubBook.images,
+                hasProgressSaved = readerData != null,
+                lastChapterIndex = readerData?.lastChapterIndex ?: 0,
+                lastChapterOffset = readerData?.lastChapterOffset ?: 0,
+                currentChapterIndex = readerData?.lastChapterIndex ?: 0
+            )
+            onLoaded(state.value)
             // Added some delay to avoid choppy animation.
-            if (state.shouldShowLoader) {
+            if (state.value.shouldShowLoader) {
                 delay(200L)
             }
-            state = state.copy(isLoading = false)
+            _state.value = _state.value.copy(
+                isLoading = false,
+                shouldShowLoader = false
+            )
         }
     }
 
@@ -112,10 +134,21 @@ class ReaderViewModel @Inject constructor(
             fileStream.use { fis ->
                 // parse and create epub book
                 val epubBook = epubParser.createEpubBook(fis, shouldUseToc = false)
-                state = state.copy(epubBook = epubBook)
+                _state.value = _state.value.copy(
+                    title = epubBook.title,
+                    chapters = epubBook.chapters,
+                    images = epubBook.images,
+                    hasProgressSaved = false,
+                    lastChapterIndex = 0,
+                    lastChapterOffset = 0,
+                    currentChapterIndex = 0
+                )
                 // Added some delay to avoid choppy animation.
                 delay(200L)
-                state = state.copy(isLoading = false)
+                _state.value = _state.value.copy(
+                    isLoading = false,
+                    shouldShowLoader = false
+                )
             }
         }
     }
@@ -124,15 +157,17 @@ class ReaderViewModel @Inject constructor(
         viewModelScope.launch(Dispatchers.IO) {
             val readerData = readerDao.getReaderData(libraryItemId)
             // if the user is not on last chapter, save the progress.
-            if (readerData != null && chapterIndex != state.epubBook?.chapters!!.size - 1) {
-                val newReaderData = readerData.copy(
+            if (state.value.hasProgressSaved && chapterIndex != state.value.chapters.size - 1) {
+                val newReaderData = readerData?.copy(
                     lastChapterIndex = chapterIndex,
                     lastChapterOffset = chapterOffset,
                     lastReadTime = System.currentTimeMillis()
                 )
-                newReaderData.id = readerData.id
-                readerDao.update(newReaderData)
-            } else if (chapterIndex == state.epubBook?.chapters!!.size - 1) {
+                newReaderData?.let {
+                    it.id = readerData.id
+                    readerDao.update(it)
+                }
+            } else if (chapterIndex == state.value.chapters.size - 1) {
                 // if the user has reached last chapter, delete this book
                 // from reader database instead of saving it's progress.
                 readerData?.let { readerDao.delete(it.libraryItemId) }
@@ -150,40 +185,29 @@ class ReaderViewModel @Inject constructor(
     }
 
     fun setChapterScrollPercent(percent: Float) {
-        _chapterScrolledPercent.value = percent
+        _state.value = _state.value.copy(chapterScrollPercent = percent)
     }
 
     fun setVisibleChapterIndex(index: Int) {
-        _visibleChapterIndex.value = index
+        _state.value = _state.value.copy(currentChapterIndex = index)
     }
 
     fun toggleReaderMenu() {
-        state = state.copy(showReaderMenu = !state.showReaderMenu)
+        _state.value = _state.value.copy(showReaderMenu = !state.value.showReaderMenu)
     }
 
     fun hideReaderInfo() {
-        state = state.copy(showReaderMenu = false)
+        _state.value = _state.value.copy(showReaderMenu = false)
     }
 
     fun setFontFamily(font: ReaderFont) {
         preferenceUtil.putString(PreferenceUtil.READER_FONT_STYLE_STR, font.id)
-        state = state.copy(fontFamily = font)
-    }
-
-    fun getFontFamily(): ReaderFont {
-        return ReaderFont.getAllFonts().find {
-            it.id == preferenceUtil.getString(
-                PreferenceUtil.READER_FONT_STYLE_STR,
-                ReaderFont.System.id
-            )
-        }!!
+        _state.value = _state.value.copy(fontFamily = font)
     }
 
     fun setFontSize(newValue: Int) {
         preferenceUtil.putInt(PreferenceUtil.READER_FONT_SIZE_INT, newValue)
-        state = state.copy(fontSize = newValue)
+        _state.value = _state.value.copy(fontSize = newValue)
     }
-
-    private fun getFontSize() = preferenceUtil.getInt(PreferenceUtil.READER_FONT_SIZE_INT, 100)
 
 }
